Update roborock install instractions
I'm spend a lot of time installing Valetudo on my Xiaomi Gen 1.
At last I've managed to flash the new software to the robot,
but it turned out that Valetudo is not starting.

After some investigation I have found the source of the problem.
The script builder_vacuum.sh parameter --valetudo-path=
should get not only the binary, but also some files
valetudo.conf, hosts and rc.local

I've asked the author of  builder_vacuum.sh and he said that the
best way is just to pass the whole git repo. But the compiled file
is also needed.

By the way, the script builder_vacuum.sh is already updated to
show errors:
https://github.com/zvldz/vacuum/commit/b27d536448df4b3cff798a9c6ad36f6eb387ae79
I