Block ott.io.mi.com requests (those are for TCP cloud connect which isn’t implemented).
While at it, remove duplicate /gslb implementation from miio-specific http server.

It cannot be reached from any other port but 80, so is of no use unless
we use iptables or patch the binaries (both of which we don’t want).

This makes our life a bit easier as there’s just one implementation
to maintain.