Setup S4 Max to handle mi iot messages
For the properties and events that have been observed, parse out the
messages and forward the required parameters to the parseAndUpdateState
method to handle updating the state of the robot.  With these changes in
place, I can complete a cleaning successfully as well as trigger an
error by lifting during cleaning and not see any errors in the Valetudo
log.