feat(vendor.roborock): Add support for the S4 Max (#699)
* Add functionality to run S4 Max

Create the configuration for the S4 Max.  This robot adds the multimap
persistence functionality to the S4 (basically the same as the S5 Max
without the mop).  This has been tested on my current S4 Max and works
correctly.

* Setup S4 Max to handle mi iot messages

For the properties and events that have been observed, parse out the
messages and forward the required parameters to the parseAndUpdateState
method to handle updating the state of the robot.  With these changes in
place, I can complete a cleaning successfully as well as trigger an
error by lifting during cleaning and not see any errors in the Valetudo
log.