viomi: Rework basic control logic
The basic control capability now remembers the last cleaning operation command
parameters and performs proper pausing/stopping.

This also means that this capability MUST be used by other capabilities that
intend to somehow control the vacuum cleaning modes, so that it gets a chance
to remember the internal state.

To do this, the capability provides some new Viomi-specific methods which
other capabilities can use instead of sending commands on their own.

The capability will additionally ensure that the mop mode is correct according
to the installed accessories, and allow the user to resume the operation.