docs: add group import, clear up import
The `groups.yaml` file doesn't seem to be imported by default and I was a bit confused why the card always said, that it could not find the group. Also, I had my own customizations already in place and needed to copy them over. This hopefully makes it clearer for other people. Additionally, this adds Syntax Highlighting on that one code block. Wdyt 👀 ?